+++ common/src/main/java/it/crystalnest/leathered_boots/CommonModLoader.java
@@ -1,6 +1,5 @@
 package it.crystalnest.leathered_boots;
 
-import it.crystalnest.leathered_boots.enchantment.EnchantmentRegistry;
 import it.crystalnest.leathered_boots.item.ItemRegistry;
 import it.crystalnest.leathered_boots.loot.CommonLootRegistry;
 import org.jetbrains.annotations.ApiStatus;
@@ -17,7 +16,6 @@ public final class CommonModLoader {
    */
   public static void init() {
     ItemRegistry.register();
-    EnchantmentRegistry.register();
     CommonLootRegistry.register();
     Constants.LEATHERED_BOOTS_DYNAMIC_DATA_PACK.register();
   }