+++ common/src/main/java/it/crystalnest/leathered_boots/Constants.java
@@ -4,19 +4,18 @@ import com.google.common.base.Suppliers;
 import it.crystalnest.cobweb.api.pack.DynamicDataPack;
 import it.crystalnest.cobweb.api.pack.DynamicTagBuilder;
 import it.crystalnest.leathered_boots.api.LeatheredBootsManager;
-import it.crystalnest.leathered_boots.item.LeatheredArmorMaterial;
+import it.crystalnest.leathered_boots.item.ItemRegistry;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.ItemTags;
-import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
+import net.minecraft.world.item.trading.ItemCost;
 import net.minecraft.world.item.trading.MerchantOffer;
 import org.jetbrains.annotations.ApiStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.Map;
-import java.util.Objects;
 import java.util.function.Supplier;
 
 /**
@@ -25,7 +24,7 @@ import java.util.function.Supplier;
 @ApiStatus.Internal
 public final class Constants {
   /**
-   * Mod id.
+   * Mod ID.
    */
   public static final String MOD_ID = "leathered_boots";
 
@@ -36,19 +35,19 @@ public final class Constants {
 
   /**
    * Leathered boots dynamic datapack.<br />
-   * Adds all leathered boots to {@link ItemTags#TRIMMABLE_ARMOR} and {@link ItemTags#FREEZE_IMMUNE_WEARABLES}.
+   * Adds all leathered boots to {@link ItemTags#DYEABLE}, {@link ItemTags#TRIMMABLE_ARMOR}, {@link ItemTags#TRIMMABLE_ARMOR} and {@link ItemTags#FOOT_ARMOR}.
    */
   public static final DynamicDataPack LEATHERED_BOOTS_DYNAMIC_DATA_PACK = (DynamicDataPack) DynamicDataPack
-    .named(new ResourceLocation(MOD_ID, "leathered_boots"))
-    .add(() -> DynamicTagBuilder.of(Registries.ITEM, ItemTags.TRIMMABLE_ARMOR, ItemTags.FREEZE_IMMUNE_WEARABLES).addElements(LeatheredBootsManager.getBoots()));
+    .named(ResourceLocation.fromNamespaceAndPath(MOD_ID, "leathered_boots"))
+    .add(() -> DynamicTagBuilder.of(Registries.ITEM, ItemTags.DYEABLE, ItemTags.TRIMMABLE_ARMOR, ItemTags.FREEZE_IMMUNE_WEARABLES, ItemTags.FOOT_ARMOR).addElements(LeatheredBootsManager.getBoots()));
 
   /**
    * Leathered boots trade offers.
    */
   public static final Supplier<Map<Integer, MerchantOffer>> LEATHERED_BOOTS_TRADES = Suppliers.memoize(() -> Map.ofEntries(
-    Map.entry(3, new MerchantOffer(new ItemStack(Items.EMERALD, 4), Objects.requireNonNull(LeatheredBootsManager.getBootsStack(Constants.MOD_ID, LeatheredArmorMaterial.LEATHERED_CHAIN)), 5, 6, 0.02F)),
-    Map.entry(4, new MerchantOffer(new ItemStack(Items.EMERALD, 9), Objects.requireNonNull(LeatheredBootsManager.getBootsStack(Constants.MOD_ID, LeatheredArmorMaterial.LEATHERED_IRON)), 3, 10, 0.02F)),
-    Map.entry(5, new MerchantOffer(new ItemStack(Items.EMERALD, 13), Objects.requireNonNull(LeatheredBootsManager.getBootsStack(Constants.MOD_ID, LeatheredArmorMaterial.LEATHERED_DIAMOND)), 1, 30, 0.02F))
+    Map.entry(3, new MerchantOffer(new ItemCost(Items.EMERALD, 4), ItemRegistry.LEATHERED_CHAIN_BOOTS.get().getDefaultInstance(), 5, 6, 0.02F)),
+    Map.entry(4, new MerchantOffer(new ItemCost(Items.EMERALD, 9), ItemRegistry.LEATHERED_IRON_BOOTS.get().getDefaultInstance(), 3, 10, 0.02F)),
+    Map.entry(5, new MerchantOffer(new ItemCost(Items.EMERALD, 13), ItemRegistry.LEATHERED_DIAMOND_BOOTS.get().getDefaultInstance(), 1, 30, 0.02F))
   ));
 
   /**