+++ common/src/main/java/it/crystalnest/leathered_boots/item/LeatheredBootsItem.java
@@ -1,23 +1,29 @@
 package it.crystalnest.leathered_boots.item;
 
-import net.minecraft.world.item.DyeableArmorItem;
+import net.minecraft.core.Holder;
+import net.minecraft.world.item.ArmorItem;
+import net.minecraft.world.item.ArmorMaterial;
 import org.jetbrains.annotations.NotNull;
 
 /**
  * Leathered Boots Item.
  */
-public class LeatheredBootsItem extends DyeableArmorItem {
+public class LeatheredBootsItem extends ArmorItem {
   /**
-   * @param armorMaterial {@link LeatheredArmorMaterial}.
-   * @param isFireResistant {@link #isFireResistant}.
+   * @param armorMaterial {@link ArmorMaterial}.
+   * @param isFireResistant whether the boots are fire-resistant.
    */
-  public LeatheredBootsItem(LeatheredArmorMaterial armorMaterial, boolean isFireResistant) {
-    super(armorMaterial, Type.BOOTS, isFireResistant ? new Properties().fireResistant() : new Properties());
+  public LeatheredBootsItem(Holder<ArmorMaterial> armorMaterial, int durabilityFactor, boolean isFireResistant) {
+    super(armorMaterial, Type.BOOTS, isFireResistant ? getProperties(durabilityFactor).fireResistant() : getProperties(durabilityFactor));
+  }
+
+  private static Properties getProperties(int durabilityFactor) {
+    return new Properties().durability(Type.BOOTS.getDurability(durabilityFactor));
   }
 
   @NotNull
   @Override
-  public LeatheredArmorMaterial getMaterial() {
-    return (LeatheredArmorMaterial) super.getMaterial();
+  public Holder<ArmorMaterial> getMaterial() {
+    return super.getMaterial();
   }
 }