+++ common/src/main/java/it/crystalnest/leathered_boots/mixin/PowderSnowBlockMixin.java
@@ -1,9 +1,11 @@
 package it.crystalnest.leathered_boots.mixin;
 
 import com.llamalad7.mixinextras.injector.ModifyReturnValue;
-import it.crystalnest.leathered_boots.enchantment.EnchantmentRegistry;
-import it.crystalnest.leathered_boots.enchantment.SoftStepEnchantment;
+import it.crystalnest.leathered_boots.Constants;
 import it.crystalnest.leathered_boots.item.LeatheredBootsItem;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
@@ -23,13 +25,16 @@ public abstract class PowderSnowBlockMixin {
    *
    * @param original original condition value.
    * @param entity {@link Entity} the entity to check.
-   * @return condition extended to take into account {@link LeatheredBootsItem} and {@link SoftStepEnchantment}.
+   * @return condition extended to take into account {@link LeatheredBootsItem} and Soft Step enchantment.
    */
   @ModifyReturnValue(method = "canEntityWalkOnPowderSnow", at = @At(value = "RETURN", ordinal = 1))
   private static boolean modifyCanEntityWalkOnPowderSnow(boolean original, Entity entity) {
     return original || entity instanceof LivingEntity livingEntity && (
       livingEntity.getItemBySlot(EquipmentSlot.FEET).getItem() instanceof LeatheredBootsItem ||
-      EnchantmentHelper.getItemEnchantmentLevel(EnchantmentRegistry.SOFT_STEP.get(), livingEntity.getItemBySlot(EquipmentSlot.FEET)) > 0
+      EnchantmentHelper.getItemEnchantmentLevel(
+        entity.level().holderLookup(Registries.ENCHANTMENT).getOrThrow(ResourceKey.create(Registries.ENCHANTMENT, ResourceLocation.fromNamespaceAndPath(Constants.MOD_ID, "soft_step"))),
+        livingEntity.getItemBySlot(EquipmentSlot.FEET)
+      ) > 0
     );
   }
 }