+++ fabric/src/main/java/it/crystalnest/leathered_boots/ClientModLoader.java
@@ -6,7 +6,7 @@ import it.crystalnest.leathered_boots.item.LeatheredBootsItem;
 import it.crystalnest.leathered_boots.platform.Services;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.rendering.v1.ColorProviderRegistry;
-import net.minecraft.world.item.DyeableLeatherItem;
+import net.minecraft.world.item.component.DyedItemColor;
 import org.jetbrains.annotations.ApiStatus;
 
 /**
@@ -16,7 +16,7 @@ import org.jetbrains.annotations.ApiStatus;
 public final class ClientModLoader implements ClientModInitializer {
   @Override
   public void onInitializeClient() {
-    ColorProviderRegistry.ITEM.register((stack, tintIndex) -> tintIndex > 0 ? -1 : ((DyeableLeatherItem) stack.getItem()).getColor(stack), LeatheredBootsManager.getBoots().toArray(LeatheredBootsItem[]::new));
+    ColorProviderRegistry.ITEM.register((stack, tintIndex) -> tintIndex > 0 ? -1 : DyedItemColor.getOrDefault(stack, DyedItemColor.LEATHER_COLOR), LeatheredBootsManager.getBoots().toArray(LeatheredBootsItem[]::new));
     if (Services.PLATFORM.isModLoaded("detailab")) {
       DetailArmorBar.register();
     }