+++ fabric/src/main/java/it/crystalnest/leathered_boots/handler/LootTableEventsHandler.java
@@ -1,20 +1,15 @@
 package it.crystalnest.leathered_boots.handler;
 
-import it.crystalnest.leathered_boots.Constants;
-import it.crystalnest.leathered_boots.api.LeatheredBootsManager;
 import it.crystalnest.leathered_boots.item.ItemRegistry;
-import it.crystalnest.leathered_boots.item.LeatheredArmorMaterial;
 import it.crystalnest.leathered_boots.item.LeatheredBootsItem;
 import it.crystalnest.leathered_boots.loot.BiomesCheck;
-import net.fabricmc.fabric.api.loot.v2.LootTableEvents;
-import net.fabricmc.fabric.api.loot.v2.LootTableSource;
+import net.fabricmc.fabric.api.loot.v3.LootTableEvents;
+import net.fabricmc.fabric.api.loot.v3.LootTableSource;
+import net.minecraft.core.HolderLookup;
 import net.minecraft.resources.ResourceKey;
-import net.minecraft.resources.ResourceLocation;
-import net.minecraft.server.packs.resources.ResourceManager;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.biome.Biome;
 import net.minecraft.world.level.biome.Biomes;
-import net.minecraft.world.level.storage.loot.LootDataManager;
 import net.minecraft.world.level.storage.loot.LootPool;
 import net.minecraft.world.level.storage.loot.LootTable;
 import net.minecraft.world.level.storage.loot.entries.LootItem;
@@ -23,6 +18,7 @@ import net.minecraft.world.level.storage.loot.predicates.LootItemRandomChanceCon
 import net.minecraft.world.level.storage.loot.providers.number.ConstantValue;
 
 import java.util.List;
+import java.util.function.Supplier;
 
 /**
  * {@link LootTableEvents} handler.
@@ -33,23 +29,22 @@ public final class LootTableEventsHandler {
   /**
    * Handles modifying Vanilla loot tables to include mod items.
    *
-   * @param resourceManager resource manager.
-   * @param lootManager loot manager.
-   * @param id loot table ID.
-   * @param builder loot table builder.
+   * @param key loot table key.
+   * @param builder builder of the loot table being loaded.
    * @param source loot table source.
+   * @param provider holder lookup provider.
    */
-  public static void handle(ResourceManager resourceManager, LootDataManager lootManager, ResourceLocation id, LootTable.Builder builder, LootTableSource source) {
-    switch (id.toString()) {
+  public static void handle(ResourceKey<LootTable> key, LootTable.Builder builder, LootTableSource source, HolderLookup.Provider provider) {
+    switch (key.location().toString()) {
       case "minecraft:chests/village/village_snowy_house" -> {
-        builder.pool(buildPool(0.2F, LeatheredArmorMaterial.LEATHERED_CHAIN));
-        builder.pool(buildPool(0.1F, LeatheredArmorMaterial.LEATHERED_IRON));
-        builder.pool(buildPool(0.05F, LeatheredArmorMaterial.LEATHERED_DIAMOND));
+        builder.pool(buildPool(0.2F, ItemRegistry.LEATHERED_CHAIN_BOOTS));
+        builder.pool(buildPool(0.1F, ItemRegistry.LEATHERED_IRON_BOOTS));
+        builder.pool(buildPool(0.05F, ItemRegistry.LEATHERED_DIAMOND_BOOTS));
       }
       case "minecraft:chests/igloo_chest" -> {
-        builder.pool(buildPool(0.2F, LeatheredArmorMaterial.LEATHERED_GOLD));
-        builder.pool(buildPool(0.1F, LeatheredArmorMaterial.LEATHERED_IRON));
-        builder.pool(buildPool(0.05F, LeatheredArmorMaterial.LEATHERED_DIAMOND));
+        builder.pool(buildPool(0.2F, ItemRegistry.LEATHERED_GOLDEN_BOOTS));
+        builder.pool(buildPool(0.1F, ItemRegistry.LEATHERED_IRON_BOOTS));
+        builder.pool(buildPool(0.05F, ItemRegistry.LEATHERED_DIAMOND_BOOTS));
         builder.pool(buildPool(0.5F, ItemRegistry.LEATHER_UPGRADE_SMITHING_TEMPLATE_ITEM.get()));
       }
       case "minecraft:chests/shipwreck_map", "minecraft:chests/shipwreck_treasure", "minecraft:chests/shipwreck_supply" -> builder.pool(buildPool(0.2F, Biomes.SNOWY_BEACH));
@@ -60,16 +55,17 @@ public final class LootTableEventsHandler {
   /**
    * Handles replacing Vanilla loot tables to include mod items.
    *
-   * @param resourceManager resource manager.
-   * @param lootManager loot manager.
-   * @param id loot table ID.
-   * @param original original loot table.
+   * @param key loot table key.
+   * @param original loot table being modified.
    * @param source loot table source.
-   * @return replaced {@link LootTable} or {@code null}.
+   * @param provider holder lookup provider.
    */
-  public static LootTable handle(ResourceManager resourceManager, LootDataManager lootManager, ResourceLocation id, LootTable original, LootTableSource source) {
-    if ("minecraft:archaeology/trail_ruins_rare".equals(id.toString())) {
-      return LootTable.lootTable().pool(LootPool.lootPool().with(original.pools.get(0).entries).with(LootItem.lootTableItem(ItemRegistry.LEATHER_UPGRADE_SMITHING_TEMPLATE_ITEM.get()).when(BiomesCheck.builder(Biomes.SNOWY_TAIGA)).build()).build()).build();
+  public static LootTable handle(ResourceKey<LootTable> key, LootTable original, LootTableSource source, HolderLookup.Provider provider) {
+    if ("minecraft:archaeology/trail_ruins_rare".equals(key.location().toString())) {
+      return LootTable
+        .lootTable()
+        .pool(LootPool.lootPool().with(original.pools.getFirst().entries).with(LootItem.lootTableItem(ItemRegistry.LEATHER_UPGRADE_SMITHING_TEMPLATE_ITEM.get()).when(BiomesCheck.builder(Biomes.SNOWY_TAIGA)).build()).build())
+        .build();
     }
     return null;
   }
@@ -91,14 +87,14 @@ public final class LootTableEventsHandler {
   }
 
   /**
-   * Builds a loot pool with the given random {@code chance} to find a {@link LeatheredBootsItem} of the given {@link LeatheredArmorMaterial}.
+   * Builds a loot pool with the given random {@code chance} to find the given {@link LeatheredBootsItem}.
    *
    * @param chance chance to spawn the item.
-   * @param armorMaterial {@link LeatheredArmorMaterial}.
+   * @param item leathered boots item.
    * @return loot pool.
    */
-  private static LootPool buildPool(float chance, LeatheredArmorMaterial armorMaterial) {
-    return buildPool(chance, LeatheredBootsManager.getBoots(Constants.MOD_ID, armorMaterial));
+  private static LootPool buildPool(float chance, Supplier<? extends Item> item) {
+    return buildPool(chance, item.get());
   }
 
   /**