+++ fabric/src/main/java/it/crystalnest/leathered_boots/platform/FabricItemHelper.java
@@ -1,12 +1,13 @@
 package it.crystalnest.leathered_boots.platform;
 
 import com.google.common.base.Suppliers;
-import it.crystalnest.leathered_boots.item.LeatheredArmorMaterial;
 import it.crystalnest.leathered_boots.item.LeatheredBootsItem;
 import it.crystalnest.leathered_boots.platform.services.ItemHelper;
 import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroup;
+import net.minecraft.core.Holder;
 import net.minecraft.core.cauldron.CauldronInteraction;
 import net.minecraft.network.chat.Component;
+import net.minecraft.world.item.ArmorMaterial;
 import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.ItemStack;
 
@@ -17,9 +18,9 @@ import java.util.function.Supplier;
  */
 public final class FabricItemHelper extends ItemHelper {
   @Override
-  public Supplier<LeatheredBootsItem> supplyItem(LeatheredArmorMaterial armorMaterial, boolean isFireResistant) {
+  public Supplier<LeatheredBootsItem> supplyItem(int durabilityFactor, boolean isFireResistant, Holder<ArmorMaterial> armorMaterial) {
     return Suppliers.memoize(() -> {
-      LeatheredBootsItem item = new LeatheredBootsItem(armorMaterial, isFireResistant);
+      LeatheredBootsItem item = new LeatheredBootsItem(armorMaterial, durabilityFactor, isFireResistant);
       CauldronInteraction.WATER.map().put(item, CauldronInteraction.DYED_ITEM);
       return item;
     });