+++ neoforge/src/main/java/it/crystalnest/leathered_boots/handler/ClientEventsHandler.java
@@ -4,16 +4,16 @@ import it.crystalnest.leathered_boots.Constants;
 import it.crystalnest.leathered_boots.api.LeatheredBootsManager;
 import it.crystalnest.leathered_boots.item.LeatheredBootsItem;
 import net.minecraft.client.color.item.ItemColor;
-import net.minecraft.world.item.DyeableLeatherItem;
+import net.minecraft.world.item.component.DyedItemColor;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.SubscribeEvent;
-import net.neoforged.fml.common.Mod;
+import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.client.event.RegisterColorHandlersEvent;
 
 /**
  * Client events handler.
  */
-@Mod.EventBusSubscriber(modid = Constants.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
+@EventBusSubscriber(modid = Constants.MOD_ID, bus = EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
 public final class ClientEventsHandler {
   private ClientEventsHandler() {}
 
@@ -24,6 +24,6 @@ public final class ClientEventsHandler {
    */
   @SubscribeEvent
   public static void handleItem(RegisterColorHandlersEvent.Item event) {
-    event.register((stack, tintIndex) -> tintIndex > 0 ? -1 : ((DyeableLeatherItem) stack.getItem()).getColor(stack), LeatheredBootsManager.getBoots().toArray(LeatheredBootsItem[]::new));
+    event.register((stack, tintIndex) -> tintIndex > 0 ? -1 : DyedItemColor.getOrDefault(stack, DyedItemColor.LEATHER_COLOR), LeatheredBootsManager.getBoots().toArray(LeatheredBootsItem[]::new));
   }
 }